Removed functionality re-used from chattr that is no longer needed
This codebase was originally derived from code that was re-used from the
chattr software project.  That project was a chat server.  One of the
concepts in the chat server was that of a client nickname (or "screen
name").  This is not needed for the onigumo server project.  Clients are
identified by a UUID and that is sufficient.